# Conflicts:
#	Source/LinqToDB.Tools/CodeModel/Types/WellKnownTypes.cs

+++ Source/LinqToDB/DataProvider/Access/AccessSchemaProviderBase.cs
@@ -24,7 +24,7 @@ namespace LinqToDB.DataProvider.Access
 
 		protected override string? GetProviderSpecificTypeNamespace() => null;
 
-		protected override Type? GetSystemType(string? dataType, string? columnType, DataTypeInfo? dataTypeInfo, long? length, int? precision, int? scale, GetSchemaOptions options)
+		protected override Type? GetSystemType(string? dataType, string? columnType, DataTypeInfo? dataTypeInfo, int? length, int? precision, int? scale, GetSchemaOptions options)
 		{
 			if (dataTypeInfo == null && dataType != null)
 			{
@@ -35,7 +35,7 @@ namespace LinqToDB.DataProvider.Access
 			return base.GetSystemType(dataType, columnType, dataTypeInfo, length, precision, scale, options);
 		}
 
-		protected override DataType GetDataType(string? dataType, string? columnType, long? length, int? prec, int? scale)
+		protected override DataType GetDataType(string? dataType, string? columnType, int? length, int? prec, int? scale)
 		{
 			return dataType?.ToLower() switch
 			{
